+++ common/src/main/java/net/mehvahdjukaar/amendments/reg/ModRegistry.java
@@ -131,7 +131,7 @@ public class ModRegistry {
                     EntityType.Builder.<MediumDragonFireball>of(MediumDragonFireball::new, MobCategory.MISC)
                             .sized(0.3125F, 0.3125F)
                             .clientTrackingRange(16)
-                            .fireImmune()
+                           // .fireImmune() //same as vanilla
                             .updateInterval(5));
 
     public static final Supplier<EntityType<MediumFireball>> MEDIUM_FIREBALL =
@@ -139,7 +139,7 @@ public class ModRegistry {
                     EntityType.Builder.<MediumFireball>of(MediumFireball::new, MobCategory.MISC)
                             .sized(0.3125F, 0.3125F)
                             .clientTrackingRange(16)
-                            .fireImmune()
+                            // .fireImmune() //same as vanilla
                             .updateInterval(5));
 
     public static final Supplier<Item> DRAGON_CHARGE = regItem(DRAGON_CHARGE_NAME,